requirements.txt
streamlit
psycopg2-binary
pandas
folium
streamlit-folium
